App to receive token
