+++ frontend/src/pages/Dashboard.tsx
@@ -1,29 +1,11 @@
-import React, { useEffect, useState } from 'react';
-import { Link } from 'react-router-dom';
+import React from 'react';
 
 const Dashboard = () => {
-  const [loggedIn, setLoggedIn] = useState<boolean>(false);
-
-  useEffect(() => {
-    const Token = window.localStorage.getItem('token');
-    if (Token !== '') {
-      setLoggedIn(true);
-    }
-  }, [setLoggedIn]);
-
   return (
-    <>
-      {loggedIn ? (
-        <div>
-          <p>Dashboard</p>
-          <p>Successfully logged in. Your test results will be displayed</p>
-        </div>
-      ) : (
-        <div>
-          <Link to="/login">Please login with your GitHub account first</Link>
-        </div>
-      )}
-    </>
+    <div>
+      <p>Logged in successfully</p>
+      <p>Your test result will be displayed</p>
+    </div>
   );
 };
